Keep working on crack method
