+++ test/test_helper.rb
@@ -9,6 +9,6 @@ require './lib/key'
 require './lib/enigma'
 require './lib/offset'
 require './lib/shift'
-require './lib/helper'
+require './lib/enigma_helper'
 require 'pry'
 require 'mocha/minitest'
